서비스 삭제, 메소드 객체로 이동
